+++ outputs/java/changed/COUNT_TRAILING_ZEROES_FACTORIAL_NUMBER.java
@@ -21,13 +21,13 @@ static int f_gold ( int n ) {
 
 //
 public static int f_filled ( long n ) {
-  long count = 0 ;
-  long i = 5 ;
-  while ( ( n / i >= 1 ) && ( n % i == 0 ) ) {
-    count += ( int ) ( n / i ) ;
-    i *= 5 ;
-  }
-  return ( int ) count ;
+    int count = 0 ;
+    int i = 5 ;
+    while ( ( n / i >= 1 ) && ( n % i == 0 ) ) {
+        count += ( int ) ( n / i ) ;
+        i *= 5 ;
+    }
+    return ( int ) count ;
 }
 
 public static void main(String args[]) {
