+++ outputs/java/changed/SUM_DIVISORS_1_N_1.java
@@ -21,13 +21,11 @@ static int f_gold ( int n ) {
 
 //
 public static int f_filled ( int n ) {
-  double sum = 0 ;
-  for ( int i = 1 ;
-  i <= n ;
-  i ++ ) {
-    sum += ( int ) ( n / i ) * i ;
-  }
-  return ( int ) sum ;
+    int sum = 0 ;
+    for ( int i = 1 ;  i <= n ;  i ++ ) {
+        sum += ( int ) ( n / i ) * i ;
+    }
+    return ( int ) sum ;
 }
 
 public static void main(String args[]) {
